Change models

+++ Meguri/Models/CommentImage.cs
@@ -6,12 +6,12 @@ namespace Meguri.Models {
     [Table("CommentImages")]
     [Index(nameof(CommentId), nameof(ImageId), IsUnique = true)]
     public class CommentImage {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int CommentId { get; set; }
+        public long CommentId { get; set; }
         public Comment Comment { get; set; } = null!;
 
-        public int ImageId { get; set; }
+        public long ImageId { get; set; }
         public Image Image { get; set; } = null!;
 
         /// 画像の表示順（0, 1, 2...）
